+++ lib/Models/company.dart
@@ -1,0 +1,9 @@
+/// Represents the company of the current
+class Company {
+  /// The name of this company
+  final String name;
+  /// The logo of this company
+  final String logo;
+
+  Company(this.name, {this.logo = ""});
+}