+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+## [0.0.3] - Jan. 22, 2021.
+
+* Getting the company informations on authentication
+* Now using sandbox in example
+
 ## [0.0.2] - Jan. 22, 2021.
 
 * Formatting the files according to Dartfmt
 
 ## [0.0.1] - Jan. 21, 2021.
 
-* Basic OAuth with Quickbooks Online.
-* SQL Queries on the Quickbooks engine.
+* Basic OAuth with Quickbooks Online
+* SQL Queries on the Quickbooks engine